Overrode `dired-open-xdg' to fix the spawned process getting killed
The `dired-open-xdg' provided by the dired-open package has an issue where if
the process that needs to be open takes longer than the xdg-open call emacs kill
the process. Using `call-process' instead of `start-process' solves the issue
since `start-process' instantiates the child as a subprocess which gets kill
once the parent finishes. Meanwhile, `call-process' spawns the child as a
separate process that will not get killed once the parent finishes.